+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+### 1.0.1
+* adds error checking and error messages.
+* remove all HTML content from the modal on hide.
+
 ### 1.0.0
 * initial release.